+++ R/Data_Functions.R
@@ -2689,19 +2689,6 @@ applyMP <- function(Data, MPs = NA, reps = 100, nsims=NA, silent=FALSE, parallel
   for (mp in 1:nMPs) {
     if (!silent)  message(MPs[mp])
     
-    #mp_ns <- find(MPs[mp])
-    #dlmmp <- grepl('DLMtool', mp_ns)
-    ## don't run DLMtool MPs in parallel (slower if you do)
-    #if (length(dlmmp)<1) dlmmp <- FALSE
-    #msemmp <- grepl('MSEtool', mp_ns)
-    ## don't run MSEtool MPs in parallel (slower if you do)
-    #if (length(msemmp)<1) msemmp <- FALSE
-    #
-    ## exceptions
-    #if (MPs[mp] %in% c('LBSPR', 'LBSPR_MLL')) dlmmp <- FALSE
-    #
-    #if (dlmmp |msemmp) runParallel <- FALSE
-    
     if (runParallel && parallel[mp]) {
       temp <- try(snowfall::sfLapply(1:nsims, MPs[mp], Data = Data, reps = reps), silent=TRUE)
     } else {