+++ R/runMSE.R
@@ -23,8 +23,8 @@ Simulate <- function(OM=MSEtool::testOM, parallel=FALSE, silent=FALSE) {
   }
 
   # ---- Set up parallel processing ----
-  ncpus <- set_parallel(any(parallel))
-
+  ncpus <- set_parallel(any(unlist(parallel)))
+  
   set.seed(OM@seed) # set seed for reproducibility
   nsim <- OM@nsim # number of simulations
   nyears <- OM@nyears # number of historical years
@@ -1082,14 +1082,6 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
   if (class(Hist) !='Hist')
     stop('Must provide an object of class `Hist`')
   
-  if (length(parallel) < length(MPs)) {
-    if (length(parallel) > 1) {
-      stop("length(parallel) must be equal to length(MPs)")
-    } else {
-      parallel <- rep(parallel, length(MPs))
-    }
-  }
-
   OM <- Hist@OM
   set.seed(OM@seed) # set seed for reproducibility
   nsim <- OM@nsim # number of simulations
@@ -1106,17 +1098,44 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
   if (checkMPs)
     MPs <- CheckMPs(MPs=MPs, silent=silent)
 
+  # ---- Set up parallel processing of MPs ---
+  runparallel <- FALSE 
+  if (any(parallel==TRUE)) runparallel <- TRUE
+  if (class(parallel)=='list') runparallel <- TRUE
+
   nMP <- length(MPs)  # the total number of methods used
   if (nMP < 1) stop("No valid MPs found", call.=FALSE)
 
   isrunning <- snowfall::sfIsRunning()
-  if (all(!parallel) & isrunning) snowfall::sfStop()
+  if (!runparallel & isrunning) snowfall::sfStop()
 
-  if (any(parallel)) {
+  if (runparallel) {
     if (!isrunning) setup()
     Export_customMPs(MPs)
   }
+  
+  parallel_in <- parallel
+  if (runparallel) parallel <- rep(TRUE, nMP)
+  if (!runparallel) parallel <- rep(FALSE, nMP)
+  
+  # Don't run MSEtool MPs in parallel
+  mp_ns <- sapply(MPs, find)
+  msemmp <- grep('MSEtool', mp_ns)
+  parallel[msemmp] <- FALSE
+  
+  # Don't run DLMtool MPs in parallel except LBSPR
+  dlmmp <- grep('DLMtool', mp_ns)
+  parallel[dlmmp] <- FALSE
+  parallel[grep('LBSPR', MPs)] <- TRUE
 
+  
+  # Manually specified MPs
+  if (class(parallel_in)=='list') {
+    parallel_in <- data.frame(parallel_in)
+    par_mps <- names(parallel_in)
+    parallel[match(par_mps, MPs)] <- unlist(parallel_in[1,])
+  }  
+  
   # ---- Set Management Interval for each MP ----
   if (length(interval) != nMP) interval <- rep(interval, nMP)[1:nMP]
   if (!all(interval == interval[1])) {
@@ -1758,7 +1777,7 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
 #' @param Hist Should model stop after historical simulations? Returns an object of
 #' class 'Hist' containing all historical data
 #' @param silent Should messages be printed out to the console?
-#' @param parallel Logical. Should the MSE be run using parallel processing? Can be a vector of length(MPs)
+#' @param parallel Logical. Should the MSE be run using parallel processing? See Details for more information. 
 #' @param extended Logical. Return extended projection results?
 #' if TRUE, `MSE@Misc$extended` is a named list with extended data
 #' (including historical and projection by area), and extended version of `MSE@Hist`
@@ -1767,6 +1786,13 @@ Project <- function (Hist=NULL, MPs=NA, parallel=FALSE, silent=FALSE,
 #'
 #' @describeIn runMSE Run the Historical Simulations and Forward Projections
 #'  from an object of class `OM
+#' @details 
+#' ## Running in Parallel
+#' For simple MPs, running in parallel can actually lead to an increase in computation time, due to the overhead in sending the 
+#' information over to the cores. Consequently, the data-limied MPs in DLMtool and the reference MPs in MSEtool are not run using parallel processing. 
+#' All other MPs, including custom MPs, will be run in parallel if argument is TRUE.
+#' To individually control which MPs run in parallel, `parallel` can be a named list of logical values, e.g., `parallel=list(AvC=TRUE)`.
+#' 
 #'
 #' @return Functions return objects of class \linkS4class{Hist} or \linkS4class{MSE}
 #' \itemize{
@@ -1797,16 +1823,6 @@ runMSE <- function(OM=MSEtool::testOM, MPs = NA, Hist=FALSE, silent=FALSE,
   } else {
     stop("You must specify an operating model")
   }
-  
-  # check parallel
-  if (length(parallel) < length(MPs)) {
-    if (length(parallel) > 1) {
-      stop("length(parallel) must be equal to length(MPs)")
-    } else {
-      parallel <- rep(parallel, length(MPs))
-    }
-  }
-  
 
   # check MPs
   if (checkMPs & !Hist)